Breakdown language switcher component
